file

+++ TeamProject_HONJAL/Honjal/src/main/java/com/honjal/honjal/model/ContentDTO.java
@@ -18,7 +18,7 @@ import lombok.ToString;
 public class ContentDTO {
 
 	private Long content_num;		// 글번호
-	private Integer member_num;		// 작성자 번호
+	private Long member_num;		// 작성자 번호
 	private String member_nname;	// 작성자 닉네임
 	private String board_code;		// 게시판코드
 	private String content_date;	// 작성날짜
